# Api/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/pos/v0.1/items/{itemKey}": {
      "delete": {
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "itemKey",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/items/batch/{itemKey}": {
      "get": {
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "itemKey",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/pos/v0.1/items/batch": {
      "post": {
        "tags": [
          "Items"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddItemBatchDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Items"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateItemBatchDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AddItemBatchDTO": {
      "required": [
        "itemKey"
      ],
      "type": "object",
      "properties": {
        "itemKey": {
          "format": "int32",
          "type": "integer"
        },
        "batchNo": {
          "type": "string"
        },
        "expirDt": {
          "format": "date-time",
          "type": "string"
        },
        "costPrice": {
          "format": "float",
          "type": "number"
        },
        "salePrice": {
          "format": "float",
          "type": "number"
        },
        "quantity": {
          "format": "float",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "UpdateItemBatchDTO": {
      "required": [
        "itemBatchKey",
        "itemKey"
      ],
      "type": "object",
      "properties": {
        "itemBatchKey": {
          "format": "int32",
          "type": "integer"
        },
        "itemKey": {
          "format": "int32",
          "type": "integer"
        },
        "batchNo": {
          "type": "string"
        },
        "expirDt": {
          "format": "date-time",
          "type": "string"
        },
        "costPrice": {
          "format": "float",
          "type": "number"
        },
        "salePrice": {
          "format": "float",
          "type": "number"
        },
        "quantity": {
          "format": "float",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
  }
}